[EASY] Extract hpa_central component from hpa source file
